+++ cli/src/complete.rs
@@ -755,6 +755,43 @@ pub fn leaf_config_key_value(current: &std::ffi::OsStr) -> Vec<CompletionCandida
     }
 }
 
+pub fn config_keys_to_unset() -> Vec<CompletionCandidate> {
+    let Ok(config_level_flag) = std::env::args()
+        .filter(|arg| matches!(arg.as_str(), "--user" | "--repo" | "--workspace"))
+        .at_most_one()
+    else {
+        return Vec::new();
+    };
+
+    with_jj(|jj, _| {
+        const TEMPLATE: &str = r#"name ++ "\t" ++ source ++ "\t" ++ stringify(value).replace(regex:'\n\s*', " ") ++ "\n""#;
+        let list_output = jj
+            .build()
+            .args(["config", "list"])
+            // Only suggest unsetting overridden config options if the corresponding level is
+            // already specified.
+            .args(
+                config_level_flag
+                    .is_some()
+                    .then_some("--include-overridden"),
+            )
+            .args(config_level_flag)
+            .args(["--template", TEMPLATE])
+            .output()
+            .map_err(user_error)?;
+        Ok(String::from_utf8_lossy(&list_output.stdout)
+            .lines()
+            .filter_map(|line| line.split('\t').collect_tuple())
+            .filter(|(_, source, _)| matches!(*source, "user" | "repo" | "workspace"))
+            .map(|(name, source, value)| {
+                CompletionCandidate::new(name)
+                    .tag(Some(source.to_string().into()))
+                    .help(Some(format!("{source}: {value}").into()))
+            })
+            .collect())
+    })
+}
+
 pub fn branch_name_equals_any_revision(current: &std::ffi::OsStr) -> Vec<CompletionCandidate> {
     let Some(current) = current.to_str() else {
         return Vec::new();